+++ mjolnir/core/VerletList.hpp
@@ -132,10 +132,10 @@ void VerletList<traitsT, potentialT>::make(neighbor_list_type& neighbors,
 
 
 #ifdef MJOLNIR_SEPARATE_BUILD
-#include <mjolnir/potential/global/DebyeHuckelPotential.hpp>
-#include <mjolnir/potential/global/ExcludedVolumePotential.hpp>
-#include <mjolnir/potential/global/LennardJonesPotential.hpp>
-#include <mjolnir/potential/global/UniformLennardJonesPotential.hpp>
+#include <mjolnir/forcefield/global/DebyeHuckelPotential.hpp>
+#include <mjolnir/forcefield/global/ExcludedVolumePotential.hpp>
+#include <mjolnir/forcefield/global/LennardJonesPotential.hpp>
+#include <mjolnir/forcefield/global/UniformLennardJonesPotential.hpp>
 
 namespace mjolnir
 {
